consolidated requisite types
